requirements.txt
Flask>=2.2,<3.1
Flask-SQLAlchemy>=3.0
python-dotenv>=0.20.0
mysqlclient>=2.1
boto3>=1.28
